Updated `formatDataSize` to use `Long` for byte size, adjusted usages, and enhanced test cases

+++ src/main/kotlin/com/vikbytes/ResultFormatter.kt
@@ -163,11 +163,11 @@ object ResultFormatter {
         }
     }
 
-    fun formatDataSize(bytes: Int): String {
+    fun formatDataSize(bytes: Long): String {
         return when {
-            bytes < 1024 -> "$bytes B"
-            bytes < 1024 * 1024 -> String.format("%.2f KB", bytes / 1024.0)
-            bytes < 1024 * 1024 * 1024 -> String.format("%.2f MB", bytes / (1024.0 * 1024.0))
+            bytes < 1024L -> "$bytes B"
+            bytes < 1024L * 1024 -> String.format("%.2f KB", bytes / 1024.0)
+            bytes < 1024L * 1024 * 1024 -> String.format("%.2f MB", bytes / (1024.0 * 1024.0))
             else -> String.format("%.2f GB", bytes / (1024.0 * 1024.0 * 1024.0))
         }
     }
